Refactor all graphql-ws related classes to graphql-apollo-ws naming, to prepare for new graphql-ws protocol implementation.
closes #302
