+++ buildSrc/src/main/groovy/io.micronaut.build.internal.graphql-module.gradle
@@ -2,3 +2,10 @@ plugins {
     id 'io.micronaut.build.internal.graphql-base'
     id 'io.micronaut.build.internal.module'
 }
+
+// TODO: Temporary disabled binary compatibility check
+micronautBuild {
+    binaryCompatibility {
+        enabled = false
+    }
+}
