+++ graphql/src/main/java/io/micronaut/configuration/graphql/apollo/ws/GraphQLWsConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.configuration.graphql.ws;
+package io.micronaut.configuration.graphql.apollo.ws;
 
 import io.micronaut.configuration.graphql.GraphQLConfiguration;
 import io.micronaut.context.annotation.ConfigurationProperties;
