+++ graphql/src/main/java/io/micronaut/configuration/graphql/apollo/ws/RequiresGraphQLWs.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.configuration.graphql.ws;
+package io.micronaut.configuration.graphql.apollo.ws;
 
 import graphql.GraphQL;
 import io.micronaut.context.annotation.Requires;
