+++ graphql/src/main/java/io/micronaut/configuration/graphql/apollo/ws/package-info.java
@@ -21,6 +21,6 @@
  */
 @Configuration
 @RequiresGraphQLWs
-package io.micronaut.configuration.graphql.ws;
+package io.micronaut.configuration.graphql.apollo.ws;
 
 import io.micronaut.context.annotation.Configuration;
